bookstore/pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "api bookstore"
authors = ["Fernando-Psy <fcesarjunior@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
